requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic==2.11.7
selenium==4.35.0
pandas==2.3.1
openpyxl==3.1.5
reportlab==4.4.3
python-multipart==0.0.20
stripe==11.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosmtplib==3.0.2
email-validator==2.2.0